backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
pymysql
python-jose[cryptography]==3.3.0
